requirements.txt
psycopg2-binary==2.9.9
pyyaml==6.0.1